[embeddable] folder cleanup (#205219)
Clean up embeddable folder structure.

PR also splits `COMMON_EMBEDDABLE_GROUPING` into
`ADD_PANEL_ANNOTATION_GROUP`, `ADD_PANEL_LEGACY_GROUP`, and
`ADD_PANEL_OTHER_GROUP` to provide a more contextually relevant name.

---------

(cherry picked from commit fd702efdd38178bd72b9a7112b9d59a971b8bc99)
