+++ src/platform/plugins/private/image_embeddable/public/actions/create_image_action.ts
@@ -10,7 +10,7 @@
 import { i18n } from '@kbn/i18n';
 import { CanAddNewPanel } from '@kbn/presentation-containers';
 import { EmbeddableApiContext } from '@kbn/presentation-publishing';
-import { COMMON_EMBEDDABLE_GROUPING } from '@kbn/embeddable-plugin/public';
+import { ADD_PANEL_ANNOTATION_GROUP } from '@kbn/embeddable-plugin/public';
 import { IncompatibleActionError, ADD_PANEL_TRIGGER } from '@kbn/ui-actions-plugin/public';
 import {
   ADD_IMAGE_EMBEDDABLE_ACTION_ID,
@@ -48,7 +48,7 @@ export const registerCreateImageAction = () => {
         // swallow the rejection, since this just means the user closed without saving
       }
     },
-    grouping: [COMMON_EMBEDDABLE_GROUPING.annotation],
+    grouping: [ADD_PANEL_ANNOTATION_GROUP],
     getDisplayName: () =>
       i18n.translate('imageEmbeddable.imageEmbeddableFactory.displayName', {
         defaultMessage: 'Image',
