--- src/platform/plugins/shared/embeddable/public/lib/types.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-export interface Trigger {
-  id: string;
-  title?: string;
-  description?: string;
-}
-
-export interface PropertySpec {
-  displayName: string;
-  accessPath: string;
-  id: string;
-  description: string;
-  value?: string;
-}
-export { ViewMode } from '../../common/types';
-
-export interface CommonlyUsedRange {
-  from: string;
-  to: string;
-  display: string;
-}
