+++ src/platform/plugins/shared/embeddable/public/plugin.tsx
@@ -17,8 +17,8 @@ import {
 } from '@kbn/core/public';
 import { Storage } from '@kbn/kibana-utils-plugin/public';
 import { migrateToLatest } from '@kbn/kibana-utils-plugin/common';
-import { bootstrap } from './bootstrap';
-import { EmbeddableStateTransfer } from './lib/state_transfer';
+import { registerTriggers } from './ui_actions/register_triggers';
+import { EmbeddableStateTransfer } from './state_transfer';
 import { EmbeddableStateWithType, CommonEmbeddableStartContract } from '../common/types';
 import {
   getExtractFunction,
@@ -47,7 +47,7 @@ export class EmbeddablePublicPlugin implements Plugin<EmbeddableSetup, Embeddabl
   constructor(initializerContext: PluginInitializerContext) {}
 
   public setup(core: CoreSetup, { uiActions }: EmbeddableSetupDependencies) {
-    bootstrap(uiActions);
+    registerTriggers(uiActions);
 
     return {
       registerReactEmbeddableFactory,
