+++ src/platform/plugins/shared/embeddable/public/react_embeddable_system/index.ts
@@ -7,6 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+export { PanelIncompatibleError } from './panel_incompatible_error';
+export { PanelNotFoundError } from './panel_not_found_error';
 export { registerReactEmbeddableFactory } from './react_embeddable_registry';
 export { ReactEmbeddableRenderer } from './react_embeddable_renderer';
 export type { DefaultEmbeddableApi, ReactEmbeddableFactory } from './types';
