+++ src/platform/plugins/shared/embeddable/public/state_transfer/embeddable_state_transfer.test.ts
