+++ src/platform/plugins/shared/embeddable/public/state_transfer/index.ts
