+++ src/platform/plugins/shared/embeddable/public/state_transfer/types.ts
