+++ src/platform/plugins/shared/embeddable/public/ui_actions/triggers.ts
